+++ android/src/main/kotlin/codeux/design/filepicker/file_picker_writable/FilePickerException.kt
@@ -1,0 +1,5 @@
+package codeux.design.filepicker.file_picker_writable
+
+class FilePickerException(message: String) : Exception(message) {
+
+}