+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0-rc.1 Feature complete for iOS and Android 🎉️
+
+* Show "Create file" dialog.
+* Show "Open file" dialog.
+* (re)read files using Uri identifier.
+* write new contents to user selected files.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial experiments